refactor: Split code for workspace subcommands into separate files in arguments package
