+++ internal/command/arguments/workspace_list.go
@@ -17,6 +17,7 @@ type Workspace struct {
 	ViewType ViewType
 }
 
+// WorkspaceList represent arguments specific to the `terraform workspace list` command.
 type WorkspaceList struct {
 	Workspace
 }
@@ -52,28 +53,4 @@ func ParseWorkspaceList(args []string) (*WorkspaceList, tfdiags.Diagnostics) {
 	default:
 		return &WorkspaceList{Workspace: Workspace{ViewType: ViewHuman}}, diags
 	}
-}
-
-type WorkspaceShow struct {
-	Workspace
-}
-
-func ParseWorkspaceShow(args []string) (*WorkspaceShow, tfdiags.Diagnostics) {
-	var diags tfdiags.Diagnostics
-
-	cmdFlags := defaultFlagSet("workspace show")
-
-	if err := cmdFlags.Parse(args); err != nil {
-		diags = diags.Append(tfdiags.Sourceless(
-			tfdiags.Error,
-			"Failed to parse command-line flags",
-			err.Error(),
-		))
-	}
-
-	// `workspace show` takes no positional arguments.
-	// We could add validation here to return an error when unexpected arguments are present,
-	// but this would be a breaking change as no validation was performed in this case before.
-
-	return &WorkspaceShow{Workspace: Workspace{ViewType: ViewHuman}}, diags
 }
